[refact]: Foldering & Separate Code by Files
