+++ NBC_W7-8_PhoneBook/NBC_W7-8_PhoneBook/Utility.swift
@@ -1,0 +1,27 @@
+//
+//  Utility.swift
+//  NBC_W7-8_PhoneBook
+//
+//  Created by 황석현 on 12/9/24.
+//
+
+import SwiftUI
+
+#if DEBUG
+extension UIViewController {
+    private struct Preview: UIViewControllerRepresentable {
+        let viewController: UIViewController
+        
+        func makeUIViewController(context: Context) -> UIViewController {
+            return viewController
+        }
+        
+        func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
+        }
+    }
+    
+    func toPreview() -> some View {
+        Preview(viewController: self)
+    }
+}
+#endif
