+++ NBC_W7-8_PhoneBook/NBC_W7-8_PhoneBook/ViewController/PhoneBookViewController.swift
@@ -107,22 +107,3 @@ class PhoneBookViewController: UIViewController {
 #Preview(body: {
     PhoneBookViewController().toPreview()
 })
-
-#if DEBUG
-extension UIViewController {
-    private struct Preview: UIViewControllerRepresentable {
-        let viewController: UIViewController
-        
-        func makeUIViewController(context: Context) -> UIViewController {
-            return viewController
-        }
-        
-        func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
-        }
-    }
-    
-    func toPreview() -> some View {
-        Preview(viewController: self)
-    }
-}
-#endif
